유림 커밋

+++ SixMonthProject/src/main/webapp/viewBoard.jsp
@@ -105,7 +105,9 @@
 					<a href="GoFree?page=1" class="button buttonSize">글 목록</a>
 				</div>
 				<div>
+				<% if(((MemberVO) session.getAttribute("vo")).getNick().equals(view.getNick())){ %>
 					<a href="GoUpdate?pnum=<%=view.getPnum()%>" class="button primary buttonSize">수정하기</a>
+					<%} %>
 				</div>
 			</div>
 		
@@ -120,12 +122,21 @@
                   <li>
                   <h3 class="medium"><%=cvo.getNick() %></h3>
                   </li>
+                  
                   <li>
                      <p class="medium"><%=cvo.getCdate() %></p>
                   </li>
+                  
                      <li class="align-right">
+                     
+                     <%-- 유림 : 밑에 조건문 한 줄 추가함 --%>
+					<% if(((MemberVO) session.getAttribute("vo")).getNick().equals(cvo.getNick())){ %>
                      <a href="ComDeleteService?cnum=<%=cvo.getCnum()%>&pnum=<%=view.getPnum()%>">X</a>
+                     <%}else{%>
+                    	<br><br>
+                     <% } %>
                      </li>
+                     
                   <p class="commentP"><%=cvo.getComments()%></p>
                  <%
 					}
